storage: extract segment code, multiple segment support; index
Logs are now comprised of multiple segments that are created after the
log reaches a certain size. For now, the hard-coded segment size is
16MiB which will be delegated to topic configuration in the future.

Each segment has a corresponding offset index, which every 4KiB writes a
new constant-size index entry and flushes the index.

This index performs a <= binary search to find the nearest file position
a particular message at a given offset can be found. Currently, this is
implemented with fseek semantics, but may be better off as a
pre-allocated mmap'd file in the future.