server: rpchandler: new storage API, cursor implementation
Stream endpoint now implements a basic cursor which allows catchup +
sends messages to subscribed entities as they are delivered. Cursors are
essentially server-side polling and per-session offset tracking, but
with the benefit that no polling delay occurs, the goroutine servicing
the streaming RPC is woken up when a message is delivered; the only
delay is then fetching all delivered messages since the notification.